PMM-5005: Change type of fields from int to int64 to avoid overflow

+++ agent/connectionuptime/service.go
@@ -27,8 +27,8 @@ type Service struct {
 
 	bits big.Int
 
-	windowPeriodSeconds int
-	indexLastStatus     int
+	windowPeriodSeconds int64
+	indexLastStatus     int64
 	startTime           time.Time
 	lastStatusTimestamp time.Time
 }
@@ -36,7 +36,7 @@ type Service struct {
 // NewService creates new instance of Service
 func NewService(windowPeriod time.Duration) *Service {
 	return &Service{
-		windowPeriodSeconds: int(windowPeriod.Seconds()),
+		windowPeriodSeconds: int64(windowPeriod.Seconds()),
 	}
 }
 
@@ -58,14 +58,14 @@ func (s *Service) registerConnectionStatus(timestamp time.Time, connected bool) 
 		return
 	}
 
-	secondsFromLastEvent := int(timestamp.Unix() - s.lastStatusTimestamp.Unix())
+	secondsFromLastEvent := timestamp.Unix() - s.lastStatusTimestamp.Unix()
 	for i := s.indexLastStatus + 1; i < (s.indexLastStatus + secondsFromLastEvent); i++ {
 		// set the same status to elements of previous connection status
-		s.bits.SetBit(&s.bits, i%s.windowPeriodSeconds, s.bits.Bit(s.indexLastStatus))
+		s.bits.SetBit(&s.bits, int(i%s.windowPeriodSeconds), s.bits.Bit(int(s.indexLastStatus)))
 	}
 
 	s.indexLastStatus = (s.indexLastStatus + secondsFromLastEvent) % s.windowPeriodSeconds
-	s.bits.SetBit(&s.bits, s.indexLastStatus, getBit(connected))
+	s.bits.SetBit(&s.bits, int(s.indexLastStatus), getBit(connected))
 	s.lastStatusTimestamp = timestamp
 }
 
@@ -94,16 +94,16 @@ func (s *Service) calculateConnectionUpTime(toTime time.Time) float32 {
 	return float32(connectedSeconds) / float32(totalNumOfSeconds) * 100
 }
 
-func (s *Service) getTotalNumberOfSeconds(toTime time.Time) int {
+func (s *Service) getTotalNumberOfSeconds(toTime time.Time) int64 {
 	totalNumOfSeconds := s.windowPeriodSeconds
-	diffInSecondsBetweenStartTimeAndToTime := int(toTime.Unix() - s.startTime.Unix())
+	diffInSecondsBetweenStartTimeAndToTime := toTime.Unix() - s.startTime.Unix()
 	if diffInSecondsBetweenStartTimeAndToTime < s.windowPeriodSeconds {
 		totalNumOfSeconds = diffInSecondsBetweenStartTimeAndToTime
 	}
 	return totalNumOfSeconds
 }
 
-func (s *Service) getStartIndex(size int) int {
+func (s *Service) getStartIndex(size int64) int64 {
 	startElement := s.indexLastStatus - size
 	if startElement < 0 {
 		startElement = s.windowPeriodSeconds + startElement
@@ -111,11 +111,11 @@ func (s *Service) getStartIndex(size int) int {
 	return startElement
 }
 
-func (s *Service) getNumOfConnectedSeconds(startIndex int, totalNumOfSeconds int) int {
+func (s *Service) getNumOfConnectedSeconds(startIndex int64, totalNumOfSeconds int64) int {
 	endIndex := startIndex + totalNumOfSeconds
 	connectedSeconds := 0
 	for i := startIndex; i < endIndex; i++ {
-		if s.bits.Bit(i%s.windowPeriodSeconds) == 1 {
+		if s.bits.Bit(int(i%s.windowPeriodSeconds)) == 1 {
 			connectedSeconds++
 		}
 	}
@@ -124,5 +124,5 @@ func (s *Service) getNumOfConnectedSeconds(startIndex int, totalNumOfSeconds int
 
 // fill values in the slice until toTime
 func (s *Service) fillStatusesUntil(toTime time.Time) {
-	s.registerConnectionStatus(toTime, s.bits.Bit(s.indexLastStatus) == 1)
+	s.registerConnectionStatus(toTime, s.bits.Bit(int(s.indexLastStatus)) == 1)
 }
